+++ maestral/utils/notify_linux.py
@@ -1,0 +1,128 @@
+# -*- coding: utf-8 -*-
+"""
+@author: Sam Schott  (ss2151@cam.ac.uk)
+
+(c) Sam Schott; This work is licensed under the MIT licence.
+
+This module handles desktop notifications for Maestral and supports multiple backends,
+depending on the platform. A single :class:`DesktopNotifier` instance is created for all
+all sync daemons and a :class:`MaestralDesktopNotifier` instance is created for each
+daemon individually. Notification settings such as as snoozing and levels can be modified
+through :class:`MaestralDesktopNotifier`.
+
+:constant int NONE: No desktop notifications.
+:constant int ERROR: Notifications on errors.
+:constant int SYNCISSUE: Notifications on sync issues.
+:constant int FILECHANGE: Notifications on file changes.
+
+"""
+
+# system imports
+from threading import Thread
+import asyncio
+import traceback
+
+# external imports
+from dbus_next import Variant
+from dbus_next.aio import MessageBus
+
+# local imports
+from .notify_base import Notification, DesktopNotifierBase, NotificationLevel
+
+
+class DBusDesktopNotifier(DesktopNotifierBase):
+    """DBus notification backend for Linux. This implements the
+    org.freedesktop.Notifications standard. The DBUS connection is created in a thread
+    with a running asyncio loop to handle clicked notifications."""
+
+    _to_native_urgency = {
+        NotificationLevel.Low: Variant('y', 0),
+        NotificationLevel.Normal: Variant('y', 1),
+        NotificationLevel.Critical: Variant('y', 2),
+    }
+
+    def __init__(self, app_name: str, app_id: str) -> None:
+        super().__init__(app_name, app_id)
+
+        self._loop = asyncio.new_event_loop()
+        self._thread = Thread(
+            target=self._start_loop,
+            args=(self._loop,),
+            daemon=True,
+            name='maestral-notification-aioloop'
+        )
+        self._thread.start()
+
+        asyncio.run_coroutine_threadsafe(self._init_dbus(), self._loop)
+
+    def _start_loop(self, loop: asyncio.AbstractEventLoop) -> None:
+        asyncio.set_event_loop(loop)
+        loop.run_forever()
+
+    async def _init_dbus(self):
+        self.bus = await MessageBus().connect()
+        introspection = await self.bus.introspect('org.freedesktop.Notifications',
+                                                  '/org/freedesktop/Notifications')
+        self.proxy_object = self.bus.get_proxy_object('org.freedesktop.Notifications',
+                                                      '/org/freedesktop/Notifications',
+                                                      introspection)
+        self.interface = self.proxy_object.get_interface('org.freedesktop.Notifications')
+        self.interface.on_action_invoked(self._on_action)
+
+    def send(self, notification: Notification) -> None:
+        asyncio.run_coroutine_threadsafe(self._send(notification), self._loop)
+
+    async def _send(self, notification: Notification) -> None:
+
+        nid = self._next_nid()
+        notification_to_replace = self.current_notifications.get(nid)
+
+        if notification_to_replace:
+            replace_id = notification_to_replace.identifier
+        else:
+            replace_id = 0
+
+        actions = ['default', 'default']
+
+        for button_name in notification.buttons.keys():
+            actions += [button_name, button_name]
+
+        try:
+            resp = await self.interface.call_notify(
+                self.app_name,         # app_name
+                replace_id,            # replaces_id
+                notification.icon,     # app_icon
+                notification.title,    # summary
+                notification.message,  # body
+                actions,               # actions
+                {                      # hints
+                    'urgency': self._to_native_urgency[notification.urgency]
+                },
+                -1,                    # expire_timeout (-1 = default)
+            )
+        except Exception:
+            # This may fail for several reasons: there may not be a systemd service
+            # file for 'org.freedesktop.Notifications' or the system configuration
+            # may have changed after DesktopNotifierFreedesktopDBus was initialized.
+            traceback.print_exc()
+        else:
+            notification.identifier = resp
+            self.current_notifications[nid] = notification
+
+    def _on_action(self, nid, action_key):
+
+        nid = int(nid)
+        action_key = str(action_key)
+        notification = next(iter(n for n in self.current_notifications.values() if n.identifier == nid), None)
+
+        if notification:
+            if action_key == 'default' and notification.action:
+                notification.action()
+            else:
+                callback = notification.buttons.get(action_key)
+
+                if callback:
+                    callback()
+
+
+Impl = DBusDesktopNotifier
